# Conflicts:
#	internal/localipc/stat_darwin.go

+++ internal/localipc/stat_darwin.go
@@ -5,9 +5,8 @@ package localipc
 import "golang.org/x/sys/unix"
 
 func unixStatDevice(stat unix.Stat_t) uint64 {
-	// #nosec G115 -- Darwin exposes dev_t as int32; widening preserves the
-	// kernel identifier for equality checks and does not size memory or files.
-	return uint64(stat.Dev)
+	// #nosec G115 -- dev_t is a signed 32-bit C type; preserve its raw identifier bits.
+	return uint64(uint32(stat.Dev))
 }
 
 func unixStatMode(stat unix.Stat_t) uint32 {
